archives/requirements.txt
absl-py==0.11.0
alembic==1.4.3
altair==4.1.0
annoy==1.17.0
appnope==0.1.2
argon2-cffi==20.1.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
base58==2.0.1
beautifulsoup4==4.9.3
bleach==3.2.1
blinker==1.4
blis==0.7.4
bs4==0.0.1
cachetools==4.2.0
catalogue==1.0.0
catboost==0.24.4
category-encoders==2.2.2
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cliff==3.5.0
cmaes==0.7.0
cmd2==1.4.0
colorama==0.4.4
colorlog==4.7.2
colour==0.1.5
configparser==5.0.1
confuse==1.4.0
cycler==0.10.0
cymem==2.0.5
decorator==4.4.2
defusedxml==0.6.0
docker-pycreds==0.4.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
entrypoints==0.3
feather-format==0.4.1
filelock==3.0.12
Flask==1.1.2
flatbuffers==1.12
gast==0.3.3
gensim==3.8.3
gitdb==4.0.5
GitPython==3.1.11
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
google-trans-new==1.1.9
graphviz==0.16
grpcio==1.32.0
h11==0.9.0
h2==3.2.0
h5py==2.10.0
hpack==3.0.0
hstspreload==2020.12.22
htmlmin==0.1.12
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
ImageHash==4.2.0
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
joblib==1.0.0
jpholiday==0.1.5
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lightgbm==3.1.1
llvmlite==0.35.0
lxml==4.6.2
Mako==1.1.4
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
matplotlib-venn==0.11.6
missingno==0.4.2
mistune==0.8.4
ml-metrics==0.1.4
murmurhash==1.0.5
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
networkx==2.5
nltk==3.5
notebook==6.1.6
numba==0.52.0
numpy==1.19.4
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
optuna==2.4.0
packaging==20.8
pandas==1.1.5
pandas-profiling==2.9.0
pandocfilters==1.4.3
parso==0.7.1
patsy==0.5.1
pbr==5.5.1
pexpect==4.8.0
phik==0.10.0
pickleshare==0.7.5
Pillow==8.0.1
plac==1.1.3
plotly==4.14.1
preshed==3.0.5
prettytable==0.7.2
prometheus-client==0.9.0
promise==2.3
prompt-toolkit==3.0.8
protobuf==3.14.0
psutil==5.8.0
ptyprocess==0.6.0
pyarrow==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydeck==0.5.0
Pygments==2.7.3
pyparsing==2.4.7
pyperclip==1.8.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-editor==1.0.4
python-pptx==0.6.18
pytz==2020.4
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==20.0.0
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
rfc3986==1.4.0
rsa==4.6
sacremoses==0.0.43
scikit-learn==0.24.0
scipy==1.5.4
seaborn==0.11.1
Send2Trash==1.5.0
sentencepiece==0.1.94
sentry-sdk==0.19.5
seqeval==1.2.2
shortuuid==1.0.1
simpletransformers==0.51.9
six==1.15.0
sklearn==0.0
smart-open==4.1.0
smmap==3.0.4
sniffio==1.2.0
soupsieve==2.1
spacy==2.3.5
SQLAlchemy==1.3.22
srsly==1.0.5
statsmodels==0.12.1
stevedore==3.3.0
streamlit==0.73.1
subprocess32==3.5.4
tangled-up-in-unicode==0.0.6
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
texthero==1.0.9
thinc==7.4.5
threadpoolctl==2.1.0
tokenizers==0.9.4
toml==0.10.2
toolz==0.11.1
torch==1.7.1
tornado==6.1
tqdm==4.54.1
traitlets==5.0.5
transformers==4.1.1
typing-extensions==3.7.4.3
tzlocal==2.1
Unidecode==1.1.2
urllib3==1.26.2
validators==0.18.2
visions==0.5.0
wandb==0.10.12
wasabi==0.8.0
watchdog==1.0.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wordcloud==1.8.1
wrapt==1.12.1
xfeat==0.1.1
XlsxWriter==1.3.7
